+++ crates/polars-lazy/src/tests/predicate_queries.rs
@@ -268,3 +268,36 @@ fn test_predicate_on_join_suffix_4788() -> PolarsResult<()> {
 
     Ok(())
 }
+
+#[test]
+fn test_push_join_col_predicates_to_both_sides_7247() -> PolarsResult<()> {
+    let df1 = df! {
+        "a" => ["a1", "a2"],
+        "b" => ["b1", "b2"],
+    }?;
+    let df2 = df! {
+        "a" => ["a1", "a1", "a2"],
+        "b2" => ["b1", "b1", "b2"],
+        "c" => ["a1", "c", "a2"]
+    }?;
+    let df = df1.lazy().join(
+        df2.lazy(),
+        [col("a"), col("b")],
+        [col("a"), col("b2")],
+        JoinArgs::new(JoinType::Inner),
+    );
+    let q = df
+        .filter(col("a").eq(lit("a1")))
+        .filter(col("a").eq(col("c")));
+
+    predicate_at_all_scans(q.clone());
+
+    let out = q.collect()?;
+    let expected = df![
+        "a" => ["a1"],
+        "b" => ["b1"],
+        "c" => ["a1"],
+    ]?;
+    assert_eq!(out, expected);
+    Ok(())
+}
